backend/requirements.txt
Django
djangorestframework
opencv-python
Pillow